made some updates to refactor how experiences and projects get rendered on screen
