+++ src/resources/icons/Flag-map_of_Nigeria.svg.png
