+++ src/resources/icons/javascript.png
