+++ src/util/fileDownloader.jsx
@@ -1,5 +1,5 @@
 import { saveAs } from 'file-saver';
-import resume from '../icons/programming_resume.pdf';
+import resume from '../resources/docs/programming_resume.pdf';
 
 function FileDownloader() {
   const dwnldFile = () => {
