+++ src/util/photoExport.js
@@ -1,14 +1,14 @@
-import CPlus from '../icons/c.png';
-import Css from '../icons/css.png';
-import Html from '../icons/html.png';
-import Java from '../icons/java.png';
-import Js from '../icons/javascript.png';
-import Linux from '../icons/linux.png';
-import NodeJs from '../icons/node-js.png';
-import ReactJs from '../icons/reactJs.png';
-import Pedalup from '../icons/pedalup.png';
-import Artsite from '../icons/artsite.png';
-import Naturalproduct from '../icons/naturalproduct.png';
+import CPlus from '../resources/icons/c.png';
+import Css from '../resources/icons/css.png';
+import Html from '../resources/icons/html.png';
+import Java from '../resources/icons/java.png';
+import Js from '../resources/icons/javascript.png';
+import Linux from '../resources/icons/linux.png';
+import NodeJs from '../resources/icons/node-js.png';
+import ReactJs from '../resources/icons/reactJs.png';
+import Pedalup from '../resources/images/pedalup.png';
+import Artsite from '../resources/images/artsite.png';
+import Naturalproduct from '../resources/images/naturalproduct.png';
 
 export {
   CPlus,
